added readme roadmap and added name field to open mic(to tell participants that they've signed up for an open mic with that name) commented out weatherforecast and pasted its content to sms controller just to figure out how this is working. you can currently hit url/sms and retrieve weather info
